Fixed time updating issue
Moved the initialization of the Date object into the getBinaryTime function. Prior to this it was declared at the beginning of the program, meaning that each time the getBinaryTime function was called it was just building a string out of the same Date object as was used last time. Reinitializing the Date object allows the program to fetch a new time, and therefore print a new time later in the program.